♻️ Split the advanced form into 2 sub-components
